segregate batching code into batch_manager
